SIS 3/backend/requirements.txt
fastapi
uvicorn[standard]
supabase
openai
python-telegram-bot[asyncio]
python-dotenv
pydantic
httpx
python-slugify
